+++ src/index.js
@@ -38,3 +38,6 @@ export { default as Price } from './molecules/Currency';
 export { default as InfoCard } from './atoms/InfoCard';
 export { default as Dropdown } from './molecules/Dropdown';
 export { default as CounterField } from './molecules/CounterField';
+export { default as LoaderPhoto } from './atoms/LoaderPhoto';
+export { default as Photo } from './molecules/Photo';
+export { default as photoGallery } from './organisms/PhotoGallery';